Rename hostsfile directory to hosts
In Go convention, the import path should match the package name. This will make code work properly with goimports
