+++ core/hosts/utils.go
@@ -25,7 +25,7 @@ type hostname struct {
 
 // newHostlist creates a hostlist by parsing a file
 func newHostlist(data []byte) *hostlist {
-	return newHostlistString(string(data));
+	return newHostlistString(string(data))
 }
 
 func newHostlistString(data string) *hostlist {
@@ -42,13 +42,13 @@ func newHostlistString(data string) *hostlist {
 }
 
 func (h *hostname) Equal(hostnamev *hostname) bool {
-	if (h.wildcard != hostnamev.wildcard || h.ipv6 != hostnamev.ipv6) {
+	if h.wildcard != hostnamev.wildcard || h.ipv6 != hostnamev.ipv6 {
 		return false
 	}
-	if (!h.ip.Equal(hostnamev.ip)) {
+	if !h.ip.Equal(hostnamev.ip) {
 		return false
 	}
-	if (h.domain != hostnamev.domain) {
+	if h.domain != hostnamev.domain {
 		return false
 	}
 	return true
@@ -56,10 +56,10 @@ func (h *hostname) Equal(hostnamev *hostname) bool {
 
 // return first match
 func (h *hostlist) FindHost(name string) (addr net.IP) {
-	var ips []net.IP;
+	var ips []net.IP
 	ips = h.FindHosts(name)
 	if len(ips) > 0 {
-		addr = ips[0];
+		addr = ips[0]
 	}
 	return
 }
@@ -73,13 +73,13 @@ func (h *hostlist) FindHosts(name string) (addrs []net.IP) {
 	}
 
 	if len(addrs) == 0 {
-		var domain_match string;
+		var domain_match string
 		for _, hostname := range *h {
 			if hostname.wildcard == true && len(hostname.domain) < len(name) {
-				domain_match = strings.Join([]string{".", hostname.domain}, "");
+				domain_match = strings.Join([]string{".", hostname.domain}, "")
 				if name[len(name)-len(domain_match):] == domain_match {
-					var left string;
-					left = name[0:len(name)-len(domain_match)]
+					var left string
+					left = name[0 : len(name)-len(domain_match)]
 					if !strings.Contains(left, ".") {
 						addrs = append(addrs, hostname.ip)
 					}
